Use VAR__ to indicate a variable, not FIXME_
Thanks to @philovivero for remarking that FIXME is a confusing variable
naming scheme -- it suggests something wrong in the code or future
work to be done, which is not accurate here.

Phil suggested something like __CURRENT_DATE__ but for some reason
I really wanted this to start with an alphanumeric.