Rename error_text to text
This is the first step to ensuring that a text/plain
response is possible to all the important situations.

That will be important because the Sandstorm installer
needs a simple text/plain response to show to the user
who is interacting with Sandcats via the installer
(which is a shell script).