Permit hyphens in hostnames
Close #45

Also adjust the benb2 domain in the tests to be ben-b2, to ensure that we can
in fact register domains with hyphens.

Also ensure that domains can't start with hyphens, or end with hyphens, or
contain two hyphens next to each other, to avoid punycode use. (Maybe one
day we'll support actual punycode.)